+++ types.go
@@ -45,10 +45,11 @@ type ErrorSource struct {
 
 // ErrorDetails type
 type ErrorDetail struct {
-	Detail string `json:"detail"`
-	Status int    `json:"status"`
-	Title  string `json:"title"`
-	Code   int    `json:"code"`
+	Detail string         `json:"detail"`
+	Status int            `json:"status"`
+	Title  string         `json:"title"`
+	Code   int            `json:"code"`
+	Errors []ErrorMessage `json:"errors"`
 }
 
 // valueOrNA turns various types into a string if they have values
@@ -67,6 +68,12 @@ func valueOrNA[T any](prefix string, quote bool, val *T) string {
 
 // a datetime type that can be unmarshaled from UNIX epoch *or* ISO times
 type uniTime time.Time
+
+// output is ISO8601(Z) down to seconds
+func (ut *uniTime) MarshalJSON() (b []byte, e error) {
+	layout := "2006-01-02T15:04:05Z"
+	return []byte("\"" + time.Time(*ut).UTC().Format(layout) + "\""), nil
+}
 
 func (ut *uniTime) UnmarshalJSON(data []byte) error {
 	// try parsing as UNIX epoch first
@@ -104,18 +111,21 @@ func parseAPIError(resp *http.Response) error {
 		return err
 	}
 
+	r := make([]string, 0)
 	if decoded.Status != 0 {
-		r := make([]string, 0)
 		r = append(r, decoded.Title)
 		for _, e := range decoded.Errors {
 			r = append(r, e.Detail)
 		}
-		return fmt.Errorf("%d %s", decoded.Status, strings.Join(r, ", "))
 	}
 
 	if decoded.Error.Status != 0 {
-		return fmt.Errorf("%d %s", decoded.Error.Status, decoded.Error.Title)
+		r = append(r, decoded.Error.Title)
+		r = append(r, decoded.Error.Detail)
+		for _, e := range decoded.Error.Errors {
+			r = append(r, e.Detail)
+		}
 	}
 
-	return fmt.Errorf("unknown error")
+	return fmt.Errorf("%d %s %s", decoded.Status, decoded.Title, strings.Join(r, ", "))
 }